+++ android/src/main/java/dev/barooni/capacitor/calendar/models/inputs/DeleteEventInput.kt
@@ -2,9 +2,11 @@ package dev.barooni.capacitor.calendar.models.inputs
 
 import com.getcapacitor.PluginCall
 import dev.barooni.capacitor.calendar.PluginError
+import dev.barooni.capacitor.calendar.models.enums.EventSpan
 
 data class DeleteEventInput(
     private val call: PluginCall,
 ) {
     val id = call.getString("id")?.toLong() ?: throw PluginError.MissingId
+    val span: EventSpan = call.getInt("span")?.let { EventSpan.fromInt(it) } ?: EventSpan.THIS_EVENT
 }